requirements.txt
requests>=2.13,<3.0
beautifulsoup4>=4.5.0,<5.0
wcwidth>=0.1.7
urwid>=2.0.0,<3.0
pillow>=8.4.0
libsixel-python>=0.5.0
